Modularize commands to take adv of categories
